pyproject.toml
[project]
name = "weightapp"
version = "0.1.0"
description = ""
authors = [
    {name = "sasilab",email = "sasistartup@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "gradio (>=5.37.0,<6.0.0)",
    "openai (>=1.96.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "plotly (>=6.2.0,<7.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "weightapp", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
